+++ tests/testthat/test-Zoopdownloader.R
@@ -11,8 +11,8 @@ Data <- Zoopdownloader(Data_folder = tempdir(), Return_object = TRUE,
                          Save_object = FALSE, Redownload_data = TRUE,
                          Data_sets = Data_sets)
 
-No_coords2<-filter(Data$Environment, is.na(Latitude) & !str_detect(Station, "NZEZ"))%>%
-  mutate(Station=paste(Source, Station))
+No_coords2<-dplyr::filter(Data$Environment, is.na(Latitude) & !stringr::str_detect(Station, "NZEZ"))%>%
+  dplyr::mutate(Station=paste(Source, Station))
 
 test_that("Dowloaded data includes all datasets", {
   expect_setequal(unique(paste(Data$Zooplankton$Source, Data$Zooplankton$SizeClass, sep="_")), Data_sets2)
@@ -35,5 +35,10 @@ test_that("Only the expected stations are missing coordinates", {
 })
 
 test_that("Date and Datetime and displaying the same dates", {
-  expect_true(all(as_date(Data$Environment$Date)==as_date(Data$Environment$Datetime) | is.na(Data$Environment$Datetime)))
+  expect_true(all(lubridate::as_date(Data$Environment$Date)==lubridate::as_date(Data$Environment$Datetime) | is.na(Data$Environment$Datetime)))
+})
+
+test_that("Bottom depths are within reasonable limits", {
+  expect_true(all(Data$Environment$BottomDepth > 0.5 | is.na(Data$Environment$BottomDepth)))
+  expect_true(all(Data$Environment$BottomDepth < 35 | is.na(Data$Environment$BottomDepth)))
 })